Linking users->games && routes -> submit attack && email capability
There’s a bit here:

Users now have game objects (which have roll objects)

Rolls are taken in by the server(though nothing is done with them yet,
the next step is to plugin a game to a roll and catalog the roll within
the game)

You can send emails… though this seems to be a little finicky. mailgun
may want me to register a real domain.

Next steps:
-link roll objects to game objects
 — then we can start building statistics!!
-prettify
-make email system work
-make user login a little more bulletproof

I think then we’re set.